Add files via upload
The Example is for Linear Regression Algorithm. To under stand how the Linear Regression Works behind the scene when you call the model. The mathematical way of solving Linear regression and Know about the Gradient Descent work behind the Linear Regression to find the best fit Line to predict the outcome.